backend/requirements.txt
django-cors-headers
djangorestframework
psycopg